+++ sdk/loganalytics/azure-resourcemanager-loganalytics/src/main/java/com/azure/resourcemanager/loganalytics/fluent/WorkspacePurgesClient.java
@@ -21,7 +21,10 @@ public interface WorkspacePurgesClient {
      * <p>In order to manage system resources, purge requests are throttled at 50 requests per hour. You should batch
      * the execution of purge requests by sending a single command whose predicate includes all user identities that
      * require purging. Use the in operator to specify multiple identities. You should run the query prior to using for
-     * a purge request to verify that the results are expected.
+     * a purge request to verify that the results are expected. Log Analytics only supports purge operations required
+     * for compliance with GDPR. The Log Analytics product team reserves the right to reject requests for purge
+     * operations that are not for the purpose of GDPR compliance. In the event of a dispute, please create a support
+     * ticket.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param workspaceName The name of the workspace.
@@ -40,7 +43,10 @@ public interface WorkspacePurgesClient {
      * <p>In order to manage system resources, purge requests are throttled at 50 requests per hour. You should batch
      * the execution of purge requests by sending a single command whose predicate includes all user identities that
      * require purging. Use the in operator to specify multiple identities. You should run the query prior to using for
-     * a purge request to verify that the results are expected.
+     * a purge request to verify that the results are expected. Log Analytics only supports purge operations required
+     * for compliance with GDPR. The Log Analytics product team reserves the right to reject requests for purge
+     * operations that are not for the purpose of GDPR compliance. In the event of a dispute, please create a support
+     * ticket.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param workspaceName The name of the workspace.
@@ -79,7 +85,7 @@ public interface WorkspacePurgesClient {
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return status of an ongoing purge operation.
+     * @return status of an ongoing purge operation along with {@link Response}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     Response<WorkspacePurgeStatusResponseInner> getPurgeStatusWithResponse(