+++ sdk/eventhubs/azure-messaging-eventhubs/CHANGELOG.md
@@ -4,11 +4,17 @@
 
 ### Features Added
 
+- Added identifier to client. ([#22981](https://github.com/Azure/azure-sdk-for-java/issues/22981))
+
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core-amqp` to `2.6.0`.
 
 ## 5.12.1 (2022-06-10)